# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Resume Optimizer API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/users/me": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get Me",
                "operationId": "get_me_api_v1_users_me_get",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfile"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Update Me",
                "operationId": "update_me_api_v1_users_me_patch",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfile"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/documents": {
            "post": {
                "tags": [
                    "documents"
                ],
                "summary": "Upload Document",
                "operationId": "upload_document_api_v1_documents_post",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_document_api_v1_documents_post"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/documents/{document_id}": {
            "get": {
                "tags": [
                    "documents"
                ],
                "summary": "Get Document",
                "operationId": "get_document_api_v1_documents__document_id__get",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Document Id"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentDetailResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "documents"
                ],
                "summary": "Delete Document",
                "operationId": "delete_document_api_v1_documents__document_id__delete",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Document Id"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/optimizations": {
            "post": {
                "tags": [
                    "optimizations"
                ],
                "summary": "Create Optimization",
                "operationId": "create_optimization_api_v1_optimizations_post",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OptimizationCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OptimizationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "optimizations"
                ],
                "summary": "List Optimizations",
                "operationId": "list_optimizations_api_v1_optimizations_get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "default": 20,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Authorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OptimizationListItem"
                                    },
                                    "title": "Response List Optimizations Api V1 Optimizations Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Body_upload_document_api_v1_documents_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "contentMediaType": "application/octet-stream",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_document_api_v1_documents_post"
            },
            "DocumentDetailResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "content_type": {
                        "type": "string",
                        "title": "Content Type"
                    },
                    "size_bytes": {
                        "type": "string",
                        "title": "Size Bytes"
                    },
                    "download_url": {
                        "type": "string",
                        "title": "Download Url"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "content_type",
                    "size_bytes",
                    "download_url"
                ],
                "title": "DocumentDetailResponse"
            },
            "DocumentResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "content_type": {
                        "type": "string",
                        "title": "Content Type"
                    },
                    "size_bytes": {
                        "type": "string",
                        "title": "Size Bytes"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "content_type",
                    "size_bytes"
                ],
                "title": "DocumentResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "OptimizationCreateRequest": {
                "properties": {
                    "document_id": {
                        "type": "string",
                        "title": "Document Id"
                    }
                },
                "type": "object",
                "required": [
                    "document_id"
                ],
                "title": "OptimizationCreateRequest"
            },
            "OptimizationListItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "document_id": {
                        "type": "string",
                        "title": "Document Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "document_id",
                    "status"
                ],
                "title": "OptimizationListItem"
            },
            "OptimizationResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "document_id": {
                        "type": "string",
                        "title": "Document Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "result": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Result"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "document_id",
                    "status"
                ],
                "title": "OptimizationResponse"
            },
...
            "UserProfile": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "is_verified": {
                        "type": "boolean",
                        "title": "Is Verified"
                    },
                    "plan": {
                        "type": "string",
                        "title": "Plan"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "email",
                    "is_verified",
                    "plan"
                ],
                "title": "UserProfile"
            },
            "UserUpdate": {
                "properties": {
                    "email": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "email"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    }
                },
                "type": "object",
                "title": "UserUpdate"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}